@@ -28,6 +28,7 @@ import android.content.Intent;
 
 import java.util.Locale;
 
+import androidx.annotation.NonNull;
 import no.nordicsemi.android.dfu.internal.exception.DeviceDisconnectedException;
 import no.nordicsemi.android.dfu.internal.exception.DfuException;
 import no.nordicsemi.android.dfu.internal.exception.RemoteDfuException;
@@ -36,7 +37,8 @@ import no.nordicsemi.android.dfu.internal.exception.UploadAbortedException;
 import no.nordicsemi.android.error.SecureDfuError;
 
 /**
- * A base class for buttonless service implementations made for Secure and in the future for Non-Secure DFU.
+ * A base class for buttonless service implementations made for Secure and in the future for
+ * Non-Secure DFU.
  */
 /* package */ abstract class ButtonlessDfuImpl extends BaseButtonlessDfuImpl {
 
@@ -44,34 +46,39 @@ import no.nordicsemi.android.error.SecureDfuError;
 
 	private static final int OP_CODE_ENTER_BOOTLOADER_KEY = 0x01;
 	private static final int OP_CODE_RESPONSE_CODE_KEY = 0x20;
-	private static final byte[] OP_CODE_ENTER_BOOTLOADER = new byte[] {OP_CODE_ENTER_BOOTLOADER_KEY};
+	private static final byte[] OP_CODE_ENTER_BOOTLOADER = new byte[]{OP_CODE_ENTER_BOOTLOADER_KEY};
 
-	ButtonlessDfuImpl(final Intent intent, final DfuBaseService service) {
+	ButtonlessDfuImpl(@NonNull final Intent intent, @NonNull final DfuBaseService service) {
 		super(intent, service);
 	}
 
 	/**
-	 * This method should return the type of the response received from the device after sending Enable Dfu command.
-	 * Should be one of {@link #NOTIFICATIONS} or {@link #INDICATIONS}.
-	 * @return response type
+	 * This method should return the type of the response received from the device after sending
+	 * Enable Dfu command. Should be one of {@link #NOTIFICATIONS} or {@link #INDICATIONS}.
+	 *
+	 * @return Response type.
 	 */
 	protected abstract int getResponseType();
 
 	/**
 	 * Returns the buttonless characteristic.
-	 * @return the characteristic used to trigger buttonless jump to bootloader mode.
+	 *
+	 * @return The characteristic used to trigger buttonless jump to bootloader mode.
 	 */
 	protected abstract BluetoothGattCharacteristic getButtonlessDfuCharacteristic();
 
 	/**
-	 * This method should return {@code true} if the bootloader is expected to start advertising with address
-	 * incremented by 1.
-	 * @return true if the bootloader may advertise with address +1, false if it will keep the same device address.
+	 * This method should return {@code true} if the bootloader is expected to start advertising
+     * with address incremented by 1.
+	 *
+	 * @return True if the bootloader may advertise with address +1, false if it will keep
+	 * the same device address.
 	 */
 	protected abstract boolean shouldScanForBootloader();
 
 	@Override
-	public void performDfu(final Intent intent) throws DfuException, DeviceDisconnectedException, UploadAbortedException {
+	public void performDfu(@NonNull final Intent intent)
+            throws DfuException, DeviceDisconnectedException, UploadAbortedException {
 		mProgressInfo.setProgress(DfuBaseService.PROGRESS_STARTING);
 
 		// Add one second delay to avoid the traffic jam before the DFU mode is enabled
@@ -109,9 +116,10 @@ import no.nordicsemi.android.error.SecureDfuError;
 
 			byte[] response;
 			try {
-				// There may be a race condition here. The peripheral should send a notification and disconnect gracefully
-				// immediately after that, but the onConnectionStateChange event may be handled before this method ends.
-				// Also, sometimes the notification is not received at all.
+				// There may be a race condition here. The peripheral should send a notification
+                // and disconnect gracefully immediately after that, but the onConnectionStateChange
+                // event may be handled before this method ends. Also, sometimes the notification
+                // is not received at all.
 				response = readNotificationResponse();
 			} catch (final DeviceDisconnectedException e) {
 				// The device disconnect event was handled before the method finished,
@@ -132,7 +140,8 @@ import no.nordicsemi.android.error.SecureDfuError;
 				 */
 				final int status = getStatusCode(response, OP_CODE_ENTER_BOOTLOADER_KEY);
 				logi("Response received (Op Code = " + response[1] + ", Status = " + status + ")");
-				mService.sendLogBroadcast(DfuBaseService.LOG_LEVEL_APPLICATION, "Response received (Op Code = " + response[1] + ", Status = " + status + ")");
+				mService.sendLogBroadcast(DfuBaseService.LOG_LEVEL_APPLICATION,
+                        "Response received (Op Code = " + response[1] + ", Status = " + status + ")");
 				if (status != DFU_STATUS_SUCCESS)
 					throw new RemoteDfuException("Device returned error after sending Enter Bootloader", status);
 				// The device will reset so we don't have to send Disconnect signal.
@@ -152,7 +161,8 @@ import no.nordicsemi.android.error.SecureDfuError;
 		} catch (final RemoteDfuException e) {
 			final int error = DfuBaseService.ERROR_REMOTE_TYPE_SECURE_BUTTONLESS | e.getErrorNumber();
 			loge(e.getMessage());
-			mService.sendLogBroadcast(DfuBaseService.LOG_LEVEL_ERROR, String.format(Locale.US, "Remote DFU error: %s", SecureDfuError.parseButtonlessError(error)));
+			mService.sendLogBroadcast(DfuBaseService.LOG_LEVEL_ERROR, String.format(Locale.US,
+                    "Remote DFU error: %s", SecureDfuError.parseButtonlessError(error)));
 			mService.terminateConnection(gatt, error | DfuBaseService.ERROR_REMOTE_MASK);
 		}
 	}
@@ -161,13 +171,15 @@ import no.nordicsemi.android.error.SecureDfuError;
 	 * Checks whether the response received is valid and returns the status code.
 	 *
 	 * @param response the response received from the DFU device.
-	 * @param request  the expected Op Code
-	 * @return the status code
-	 * @throws UnknownResponseException if response was not valid
+	 * @param request  the expected Op Code.
+	 * @return The status code.
+	 * @throws UnknownResponseException if response was not valid.
 	 */
+	@SuppressWarnings("SameParameterValue")
 	private int getStatusCode(final byte[] response, final int request) throws UnknownResponseException {
 		if (response == null || response.length < 3 || response[0] != OP_CODE_RESPONSE_CODE_KEY || response[1] != request ||
-				(response[2] != DFU_STATUS_SUCCESS && response[2] != SecureDfuError.BUTTONLESS_ERROR_OP_CODE_NOT_SUPPORTED && response[2] != SecureDfuError.BUTTONLESS_ERROR_OPERATION_FAILED))
+				(response[2] != DFU_STATUS_SUCCESS && response[2] != SecureDfuError.BUTTONLESS_ERROR_OP_CODE_NOT_SUPPORTED
+                        && response[2] != SecureDfuError.BUTTONLESS_ERROR_OPERATION_FAILED))
 			throw new UnknownResponseException("Invalid response received", response, OP_CODE_RESPONSE_CODE_KEY, request);
 		return response[2];
 	}
