@@ -32,9 +32,15 @@ import android.bluetooth.BluetoothGattCallback;
 		}
 	}
 
-	/** Returns the final BluetoothGattCallback instance, depending on the implementation. */
+	/**
+	 * Returns the final BluetoothGattCallback instance, depending on the implementation.
+	 */
 	DfuGattCallback getGattCallback();
 
-	/** Callback invoked when bond state changes to {@link android.bluetooth.BluetoothDevice#BOND_BONDED BOND_BONDED} or {@link android.bluetooth.BluetoothDevice#BOND_NONE BOND_NONE}. */
+	/**
+	 * Callback invoked when bond state changes to
+	 * {@link android.bluetooth.BluetoothDevice#BOND_BONDED BOND_BONDED} or
+     * {@link android.bluetooth.BluetoothDevice#BOND_NONE BOND_NONE}.
+	 */
 	void onBondStateChanged(final int state);
 }
