@@ -27,24 +27,55 @@ import android.content.Intent;
 
 import java.io.InputStream;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import no.nordicsemi.android.dfu.internal.exception.DeviceDisconnectedException;
 import no.nordicsemi.android.dfu.internal.exception.DfuException;
 import no.nordicsemi.android.dfu.internal.exception.UploadAbortedException;
 
 /* package */ interface DfuService extends DfuCallback {
 
-	/** This method must return true if the device is compatible with this DFU implementation, false otherwise. */
-	boolean isClientCompatible(final Intent intent, final BluetoothGatt gatt) throws DfuException, DeviceDisconnectedException, UploadAbortedException;
+	/**
+	 * This method must return true if the device is compatible with this DFU implementation,
+	 * false otherwise.
+     *
+     * @throws DeviceDisconnectedException Thrown when the device will disconnect in the middle of
+     *                                     the transmission.
+     * @throws DfuException                Thrown if DFU error occur.
+     * @throws UploadAbortedException      Thrown if DFU operation was aborted by user.
+	 */
+	boolean isClientCompatible(@NonNull final Intent intent, @NonNull final BluetoothGatt gatt)
+			throws DfuException, DeviceDisconnectedException, UploadAbortedException;
 
 	/**
 	 * Initializes the DFU implementation and does some initial setting up.
-	 * @return true if initialization was successful and the DFU process may begin, false to finish teh DFU service
+	 *
+	 * @return True if initialization was successful and the DFU process may begin,
+	 * false to finish teh DFU service.
+	 * @throws DeviceDisconnectedException Thrown when the device will disconnect in the middle of
+	 *                                     the transmission.
+	 * @throws DfuException                Thrown if DFU error occur.
+	 * @throws UploadAbortedException      Thrown if DFU operation was aborted by user.
 	 */
-	boolean initialize(final Intent intent, final BluetoothGatt gatt, final int fileType, final InputStream firmwareStream, final InputStream initPacketStream) throws DfuException, DeviceDisconnectedException, UploadAbortedException;
+	boolean initialize(@NonNull final Intent intent, @NonNull final BluetoothGatt gatt,
+					   @FileType final int fileType,
+					   @NonNull final InputStream firmwareStream,
+                       @Nullable final InputStream initPacketStream)
+			throws DfuException, DeviceDisconnectedException, UploadAbortedException;
 
-	/** Performs the DFU process. */
-	void performDfu(final Intent intent) throws DfuException, DeviceDisconnectedException, UploadAbortedException;
+	/**
+	 * Performs the DFU process.
+	 *
+	 * @throws DeviceDisconnectedException Thrown when the device will disconnect in the middle of
+	 *                                     the transmission.
+	 * @throws DfuException                Thrown if DFU error occur.
+	 * @throws UploadAbortedException      Thrown if DFU operation was aborted by user.
+	 */
+	void performDfu(@NonNull final Intent intent)
+			throws DfuException, DeviceDisconnectedException, UploadAbortedException;
 
-	/** Releases the service. */
+	/**
+	 * Releases the service.
+	 */
 	void release();
 }
