@@ -26,9 +26,11 @@ import android.content.Intent;
 import android.os.ParcelUuid;
 import android.os.Parcelable;
 
+import androidx.annotation.NonNull;
+
 /* package */ class UuidHelper {
 
-	/* package */ static void assignCustomUuids(final Intent intent) {
+	/* package */ static void assignCustomUuids(@NonNull final Intent intent) {
 		// Added in SDK 4.3.0. Legacy DFU and Legacy bootloader share the same UUIDs.
 		Parcelable[] uuids = intent.getParcelableArrayExtra(DfuBaseService.EXTRA_CUSTOM_UUIDS_FOR_LEGACY_DFU);
 		if (uuids != null && uuids.length == 4) {
